requirements.txt
twilio 
python-dotenv
os 
sys
selenium 
